requirements.txt
aioredis==1.3.1
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.8.0
certifi==2026.1.4
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
constantly==23.10.4
cryptography==46.0.4
daphne==4.2.1
Django==5.2.9
django-cors-headers==4.9.0
django-huey==1.3.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
hiredis==3.3.0
huey==2.6.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
msgpack==1.1.2
oauthlib==3.3.1
packaging==26.0
psycopg2==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.10.1
pyOpenSSL==25.3.0
redis==7.1.0
requests==2.32.5
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.4
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.3
zope.interface==8.2
